league_ranking/requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
progressbar2==3.37.1
python-utils==2.3.0
requests==2.18.4
six==1.11.0
urllib3==1.22
